requirements.txt
discord.py
peewee
pyyaml
mutagen
pafy
PyNaCl
gtts
youtube-dl